feat: StudygroupJoin 엔티티 설정 remove: classmate 엔티티 삭제, StudygroupJoin Dto, Manager 삭제 move: classmate 폴더 -> studygroupJoin
