+++ server/edusync/src/main/java/com/codestates/edusync/study/studygroupjoin/repository/ClassmateRepository.java
@@ -1,6 +1,5 @@
-package com.codestates.edusync.study.classmate.repository;
+package com.codestates.edusync.study.studygroupjoin.repository;
 
-import com.codestates.edusync.study.classmate.entity.Classmate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ClassmateRepository extends JpaRepository<Classmate, Long> {
