--- server/edusync/src/main/java/com/codestates/edusync/study/studyjoin/dto/StudygroupJoinResponseDto.java
@@ -1,13 +1,0 @@
-package com.codestates.edusync.study.studyjoin.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@Getter
-@Setter
-public class StudygroupJoinResponseDto {
-    private Long id;
-    private String classmateName;
-}
