--- server/edusync/src/main/java/com/codestates/edusync/study/studyjoin/utils/StudygroupJoinManager.java
@@ -1,4 +1,0 @@
-package com.codestates.edusync.study.studyjoin.utils;
-
-import com.codestates.edusync.study.studygroup.entity.Studygroup;
-
